spm/requirements.txt
numpy==1.26.4
matplotlib==3.8.4
supersuit==3.9.2
gymnasium==0.28.1
pettingzoo==1.24.0
stable-baselines3==2.0.0
opencv-python
pyyaml